requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
bcrypt
watchfiles
